+++ lib/warren/app/consumer.rb
@@ -32,6 +32,14 @@ module Warren
       option :path, type: :string,
                     default: Warren::Config::Consumers::DEFAULT_PATH,
                     desc: 'The path to the consumer configuration file to generate'
+      option :delay, type: :numeric,
+                     desc: 'The delay (ms) on the delay queue. 0 to skip queue creation.'
+      # Invoked by `$ warren consumer add` adds a consumer to the `warren_consumers.yml`
+      #
+      # @param name [String, nil] Optional: Passed in from Command. The name of the consumer to create.
+      #
+      # @return [Void]
+      #
       def add(name = nil)
         say 'Adding a consumer'
         Warren::App::ConsumerAdd.invoke(self, name, options)
@@ -44,6 +52,10 @@ module Warren
       option :consumers, type: :array,
                          desc: 'The consumers to start. Defaults to all consumers',
                          banner: 'consumer_name other_consumer'
+      # Invoked by `$ warren consumer start`. Starts up the configured consumers
+      #
+      # @return [Void]
+      #
       def start
         say 'Starting consumers'
         Warren::App::ConsumerStart.invoke(self, options)