+++ lib/warren/app/consumer_start.rb
@@ -7,6 +7,17 @@ module Warren
   module App
     # Handles the initial creation of the configuration object
     class ConsumerStart
+      #
+      # Starts up a warren client process for the configured consumers.
+      #
+      # @param shell [Thor::Shell::Basic] Thor shell instance for feedback
+      # @param options [Hash] Hash of command line arguments from Thor
+      # @option options [String] :path Path to the `warren_consumers.yml `file
+      # @option options [Array<String>] :consumers Array of configured consumers to start.
+      #                                            Defaults to all consumers
+      #
+      # @return [Void]
+      #
       def self.invoke(shell, options)
         new(shell, options).invoke
       end
@@ -17,6 +28,10 @@ module Warren
         @consumers = options[:consumers]
       end
 
+      #
+      # Starts up a warren client process for the configured consumers.
+      #
+      # @return [Void]
       def invoke
         Warren::Client.new(@config, consumers: @consumers).run
       end