+++ lib/warren/handler/broadcast.rb
@@ -28,17 +28,36 @@ module Warren
           @routing_key_template = Handler.routing_key_template(routing_key_prefix)
         end
 
+        # Publishes `message` to the configured exchange
+        #
+        # @param message [#routing_key,#payload] A message should respond to routing_key and payload.
+        #                                        @see Warren::Message::Full
+        #
+        # @return [Warren::Handler::Broadcast::Channel] returns self for chaining
+        #
         def <<(message)
-          default_exchange.publish(message.payload, routing_key: key_for(message))
+          publish(message)
+        end
+
+        # Publishes `message` to `exchange` (Defaults to configured exchange)
+        #
+        # @param message [#routing_key,#payload] A message should respond to routing_key and payload.
+        #                                        @see Warren::Message::Full
+        # @param exchange [Bunny::Exchange] The exchange to publish to
+        #
+        # @return [Warren::Handler::Broadcast::Channel] returns self for chaining
+        #
+        def publish(message, exchange: configured_exchange)
+          exchange.publish(message.payload, routing_key: key_for(message), headers: message.headers)
           self
         end
 
         private
 
-        def default_exchange
+        def configured_exchange
           raise StandardError, 'No exchange configured' if @exchange_name.nil?
 
-          @default_exchange ||= exchange(@exchange_name, auto_delete: false, durable: true, type: :topic)
+          @configured_exchange ||= exchange(@exchange_name, auto_delete: false, durable: true, type: :topic)
         end
 
         def key_for(message)
@@ -108,9 +127,9 @@ module Warren
         self
       end
 
-      def new_channel
-        Channel.new(session.create_channel(nil, 1), exchange: @exchange_name,
-                                                    routing_key_prefix: @routing_key_prefix)
+      def new_channel(worker_count: 1)
+        Channel.new(session.create_channel(nil, worker_count), exchange: @exchange_name,
+                                                               routing_key_prefix: @routing_key_prefix)
       end
 
       private