+++ lib/warren/message/full.rb
@@ -10,6 +10,13 @@ module Warren
         @record = record
       end
 
+      #
+      # The routing key that will be used for the message, not including the
+      # routing_key_prefix configured in warren.yml. If {#record} responds
+      # to `routing_key` will use that instead
+      #
+      # @return [String] The routing key.
+      #
       def routing_key
         if record.respond_to?(:routing_key)
           record.routing_key
@@ -18,8 +25,21 @@ module Warren
         end
       end
 
+      #
+      # The payload of the message.
+      # @see https://github.com/intridea/multi_json
+      #
+      # @return [String] The message payload
       def payload
         MultiJson.dump(record)
+      end
+
+      #
+      # For compatibility. Returns an empty hash.
+      #
+      # @return [{}] Empty hash
+      def headers
+        {}
       end
     end
   end